Refactor Tooltip and convert story
