+++ packages/manager/src/components/BackupStatus/BackupStatus.tsx
@@ -6,7 +6,7 @@ import { Tooltip } from 'src/components/Tooltip';
 import { Typography } from 'src/components/Typography';
 import { DateTimeDisplay } from 'src/components/DateTimeDisplay';
 import { Link } from 'src/components/Link';
-import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon';
 import { makeStyles } from 'tss-react/mui';
 
 const useStyles = makeStyles<void, 'icon'>()(
