+++ packages/manager/src/components/Toggle/Toggle.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import ToggleOff from 'src/assets/icons/toggleOff.svg';
 import ToggleOn from 'src/assets/icons/toggleOn.svg';
 import Switch, { SwitchProps } from '@mui/material/Switch';
-import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon';
 
 export interface ToggleProps extends SwitchProps {
   tooltipText?: JSX.Element | string;
