--- packages/manager/src/components/TooltipIcon/TooltipIcon.stories.mdx
@@ -1,57 +1,0 @@
-import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
-
-<Meta title="Components/Tooltip/Tooltip Icon" component={TooltipIcon} />
-
-export const Template = (args) => {
-  return <TooltipIcon {...args} />;
-};
-
-# Tooltip Icon
-
-## Usage
-
-Tooltips can be attached to any active element (text fields, buttons, etc.) on a page. They provide descriptions or explanations for their paired elements. Thus, tooltips are highly contextual and specific and don’t explain the bigger picture or entire task flow.
-
-**Guidelines**
-
-- Don’t use tooltips for information that is vital to task completion.
-- Provide brief and helpful content inside the tooltip.
-- Support _both_ mouse _and_ keyboard hover.
-- Present a link to additional content if needed.
-
-<Canvas>
-  <Story
-    name="Tooltip Icon"
-    args={{
-      text: 'Hello World',
-      tooltipPosition: 'bottom',
-      status: 'info',
-    }}
-    argTypes={{
-      text: {
-        description: 'Tooltip text.',
-      },
-      status: {
-        description: 'Sets the tooltip icon SVG.',
-      },
-      leaveDelay: {
-        description: 'Delay hiding the tooltip.',
-      },
-      sxTooltipIcon: {
-        description: 'Pass specific CSS styling for the SVG icon.',
-      },
-      tooltipAnalyticsEvent: {
-        description: 'Send event analytics.',
-      },
-    }}
-  >
-    {Template.bind({})}
-  </Story>
-</Canvas>
-
-<ArgsTable
-  story="Tooltip Icon"
-  exclude={['className', 'classes', 'ref']}
-  sort="requiredFirst"
-/>
