--- packages/manager/src/components/TooltipIcon/index.ts
@@ -1,2 +1,0 @@
-import { TooltipIcon } from './TooltipIcon';
-export default TooltipIcon;
