+++ packages/manager/src/features/EntityTransfers/EntityTransfersLanding/TransferControls.tsx
@@ -6,7 +6,7 @@ import { makeStyles } from '@mui/styles';
 import { Theme } from '@mui/material/styles';
 import { Typography } from 'src/components/Typography';
 import Grid from '@mui/material/Unstable_Grid2';
-import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon';
 import { TextField } from 'src/components/TextField';
 import ConfirmTransferDialog from './ConfirmTransferDialog';
