+++ packages/manager/src/features/Linodes/LinodesCreate/AddonsPanel.tsx
@@ -10,7 +10,7 @@ import { Theme } from '@mui/material/styles';
 import { Typography } from 'src/components/Typography';
 import { Currency } from 'src/components/Currency';
 import Grid from '@mui/material/Unstable_Grid2';
-import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon';
 import { CreateTypes } from 'src/store/linodeCreate/linodeCreate.actions';
 import AttachVLAN from './AttachVLAN';
 import { privateIPRegex } from 'src/utilities/ipUtils';
