+++ packages/manager/src/features/Linodes/LinodesCreate/UserDataAccordion/UserDataAccordionHeading.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { BetaChip } from 'src/components/BetaChip/BetaChip';
 import { Link } from 'src/components/Link';
 import { Notice } from 'src/components/Notice/Notice';
-import { TooltipIcon } from 'src/components/TooltipIcon/TooltipIcon';
+import { TooltipIcon } from 'src/components/TooltipIcon';
 import { Box } from 'src/components/Box';
 import { CreateTypes } from 'src/store/linodeCreate/linodeCreate.actions';
